+++ src/core_modules/capture-core/components/ListView/Main/ListViewMain.component.js
@@ -25,6 +25,7 @@ const getStyles = (theme: Theme) => ({
     },
     topBarButtonContainer: {
         display: 'flex',
+        gap: '6px',
         alignItems: 'center',
     },
     paginationContainer: {