+++ src/core_modules/capture-core/components/OrgUnitFetcher/OrgUnitFetcher.epics.js
@@ -1,17 +1,14 @@
 // @flow
 import { ofType } from 'redux-observable';
-import { catchError, map, switchMap } from 'rxjs/operators';
-import { from, of } from 'rxjs';
+import { map } from 'rxjs/operators';
+import { getCoreOrgUnit } from '../../metadataRetrieval/coreOrgUnit';
 import { actionTypes, setCurrentOrgUnit, errorRetrievingOrgUnit } from './OrgUnitFetcher.actions';
 
-
-const orgUnitsQuery = id => ({ resource: `organisationUnits/${id}` });
-
-export const orgUnitFetcherEpic = (action$: InputObservable, _: ReduxStore, { querySingleResource }: ApiUtils,
-) => action$.pipe(
+export const orgUnitFetcherEpic = (action$: InputObservable) => action$.pipe(
     ofType(actionTypes.FETCH_ORG_UNIT),
-    switchMap(({ payload: { orgUnitId } }) => from(querySingleResource(orgUnitsQuery(orgUnitId)))
-        .pipe(map(({ id, displayName: name }) => setCurrentOrgUnit({ id, name }))))
-    ,
-    catchError(() => of(errorRetrievingOrgUnit())),
+    map(({ payload: { orgUnitId } }) => getCoreOrgUnit({
+        orgUnitId,
+        onSuccess: orgUnit => setCurrentOrgUnit(orgUnit),
+        onError: errorRetrievingOrgUnit,
+    })),
 );