+++ src/core_modules/capture-core/components/Pages/New/RegistrationDataEntry/helpers/deriveAutoGenerateEvents.js
@@ -24,7 +24,7 @@ export const deriveAutoGenerateEvents = ({
     occurredAt: string,
     programId: string,
     orgUnitId: string,
-    firstStageMetadata: ProgramStage,
+    firstStageMetadata: ?ProgramStage,
     attributeCategoryOptions: { [categoryId: string]: string } | string,
 }) => {
     // in case we have a program that does not have an incident date (occurredAt), such as Malaria case diagnosis,