+++ src/core_modules/capture-core/components/Pages/common/TEIRelationshipsWidget/RegisterTei/DataEntry/TrackedEntityInstance/index.js
@@ -1,0 +1,2 @@
+// @flow
+export { DataEntryTrackedEntityInstance } from './DataEntryTrackedEntityInstance';