+++ src/core_modules/capture-core/storageControllers/index.js
@@ -3,4 +3,5 @@ export { initAsync as initControllersAsync } from './storageControllers';
 export { closeAsync as closeControllersAsync } from './storageControllers';
 export { getMainController as getMainStorageController } from './storageControllers';
 export { getUserController as getUserStorageController } from './storageControllers';
+export { userStores } from './stores';
 export type * from './cache.types';